+++ CHANGELOG.md
@@ -5,14 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-
+## [1.5.1] - 2023-10-04
 
 ### Changed
 
 - Improve error message when it is not possible to access the data storage
 - Remove `chunk_size` parameter from configuration
 - Try to take Intake URL from an environment var first  
+
+### Fixed
+
+- Create volume storage if it doesn't exist before writing cert files
 
 ## [1.5.0] - 2023-09-19